backend/requirements.txt
asgiref==3.10.0
click==8.1.8
datetime-truncate==1.1.1
dj-database-url==3.0.1
Django==4.2.25
django-admin-charts==1.5.0
django-bower==5.2.0
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-memoize==2.3.1
django-multiselectfield==1.0.1
django-nvd3==0.10.1
djangorestframework==3.14.0
gunicorn==23.0.0
h11==0.16.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
uvicorn==0.38.0
whitenoise==6.6.0
requests==2.31.0
